feat(invite-member): add QR code invitation feature and update translations
