+++ app/(tabs)/(group)/group-details/member/inivite-member/InviteMember.translate.ts
@@ -1,7 +1,7 @@
 const TEXT_TRANSLATE_INVITE_MEMBER = {
   HEADER: "Mời thành viên",
   INVITE_BY_EMAIL: "Mời bạn bè qua Email",
-  JOIN_LINK: "Link tham gia Quỹ",
+  INVITE_BY_QR_CODE: "QR tham gia nhóm",
   MEMBER_LIST: (count: number) => `Danh sách thành viên (${count})`,
   OWNER: "Người lập quỹ (Bạn)",
   CONTRIBUTION: "Đã góp quỹ",
